优化任务创建弹窗 Fixes #15
