+++ lib/assets/translations/en-US.json
@@ -1,9 +1,27 @@
 {
-    "AppName": "My Application",
+    "AppName": "KPomodoro",
+    "Common_Confirm": "Confirm",
+    "Common_Cancel": "Cancel",
     "Home_Start": "Start",
+    "Home_Skip": "Skip",
+    "Home_Resume": "Resume",
+    "Home_Break": "Break",
     "Home_Pause": "Pause",
     "Home_Reset": "Reset",
     "Home_Settings": "Settings",
     "Home_About": "About",
-    "Home_Statistics": "Statistics"
+    "Home_Statistics": "Statistics",
+    "Setting_PomodoroDuration": "Pomodoro Duration",
+    "Setting_ShortBreakDuration": "Short Break Duration",
+    "Setting_LongBreakDuration": "Long Break Duration",
+    "Setting_LongBreakInterval": "Long Break Interval",
+    "Setting_AutoBreak": "Auto Break",
+    "Setting_AutoStart": "Auto Start",
+    "Task_PomodoroCount": "Pomodoro Count",
+    "Task_Title": "Task Title",
+    "Task_Description": "Task Description",
+    "Task_Create": "Create Task",
+    "Task_Update": "Update Task",
+    "Task_StartDate": "Start Date",
+    "Task_DueDate": "Due Date"
 }
