+++ lib/assets/translations/zh-CN.json
@@ -1,9 +1,27 @@
 {
     "AppName": "KPomodoro",
+    "Common_Confirm": "确认",
+    "Common_Cancel": "取消",
     "Home_Start": "开始",
+    "Home_Skip": "跳过",
+    "Home_Resume": "继续",
+    "Home_Break": "休息",
     "Home_Pause": "停止",
     "Home_Reset": "重置",
     "Home_Settings": "设置",
     "Home_About": "关于",
-    "Home_Statistics": "统计"
+    "Home_Statistics": "统计",
+    "Setting_PomodoroDuration": "专注时长",
+    "Setting_ShortBreakDuration": "短休息时长",
+    "Setting_LongBreakDuration": "长休息时长",
+    "Setting_LongBreakInterval": "长休息间隔",
+    "Setting_AutoBreak": "自动休息",
+    "Setting_AutoStart": "自动开始",
+    "Task_PomodoroCount": "番茄钟数量",
+    "Task_Title": "任务标题",
+    "Task_Description": "任务描述",
+    "Task_Create": "创建任务",
+    "Task_Update": "更新任务",
+    "Task_StartDate": "开始日期",
+    "Task_DueDate": "截止日期"
 }
