+++ lib/core/constants/locale_keys.dart
@@ -1,6 +1,10 @@
 /// 多语言keys
 final class LocaleKeys {
   static const String AppName = "AppName";
+
+  static const String Common_Confirm = "Common_Confirm";
+  static const String Common_Cancel = "Common_Cancel";
+
   static const String Home_Start = "Home_Start";
   static const String Home_Skip = "Home_Skip";
   static const String Home_Resume = "Home_Resume";
@@ -14,4 +18,12 @@ final class LocaleKeys {
   static const String Setting_LongBreakInterval = "Setting_LongBreakInterval";
   static const String Setting_AutoBreak = "Setting_AutoBreak";
   static const String Setting_AutoStart = "Setting_AutoStart";
+
+  static const String Task_PomodoroCount = "Task_PomodoroCount";
+  static const String Task_Title = "Task_Title";
+  static const String Task_Description = "Task_Description";
+  static const String Task_Create = "Task_Create";
+  static const String Task_Update = "Task_Update";
+  static const String Task_StartDate = "Task_StartDate";
+  static const String Task_DueDate = "Task_DueDate";
 }
